python-backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.7.1 # Updated pydantic to be compatible with pydantic-settings 2.3.0
pydantic-settings==2.3.0
python-dotenv==1.0.0
numpy==1.26.1
pandas==2.1.2
scikit-learn==1.3.2
matplotlib==3.7.1
httpx==0.25.1
requests==2.30.0
aiohttp==3.8.4
aiofiles==23.1.0
sqlalchemy==2.0.23
aiosqlite==0.19.0
PyYAML==6.0.1
python-multipart==0.0.6
tiktoken==0.5.1
faker==18.9.0
openai==1.2.0
anthropic==0.5.0
google-generativeai==0.3.1
Pillow==10.3.0 # For image processing, e.g., by GeminiProvider

email-validator

pytest
